+++ src/lib.rs
@@ -2,18 +2,17 @@ mod printer;
 mod reader;
 mod types;
 
-use crate::types::Ast;
+use crate::types::Type;
 
-
-fn read(input: &str) -> Ast {
+fn read(input: &str) -> Type {
     reader::read_str(input)
 }
 
-fn eval(ast: Ast) -> Ast {
+fn eval(ast: Type) -> Type {
     ast
 }
 
-fn print(ast: Ast) -> String {
+fn print(ast: Type) -> String {
     printer::pr_str(ast)
 }