+++ src/types.rs
@@ -7,19 +7,3 @@ pub enum Type {
     List(Vec<Box<Type>>),
 }
 
-#[derive(Debug,PartialEq)]
-pub struct Ast {
-    root: Type,
-}
-
-impl Ast {
-    pub fn new(root: Type) -> Ast {
-        Ast {
-            root: root
-        }
-    }
-
-    pub fn root(&self) -> &Type {
-        &self.root
-    }
-}